# dev/core#4567 Add upgrade script to mailing-queue-table alter