--- a/CRM/Upgrade/Incremental/php/FiveSixtySeven.php
+++ b/CRM/Upgrade/Incremental/php/FiveSixtySeven.php
@@ -85,6 +85,11 @@ class CRM_Upgrade_Incremental_php_FiveSixtySeven extends CRM_Upgrade_Incremental
       FOREIGN KEY (`job_id`)
       REFERENCES `civicrm_mailing_job`(`id`) ON DELETE SET NULL
     ', [], FALSE, FALSE, FALSE, FALSE);
+      CRM_Core_DAO::executeQuery('
+UPDATE  civicrm_mailing_event_queue q
+INNER JOIN civicrm_mailing_job job ON job.id = q.job_id
+SET q.mailing_id = job.mailing_id, q.is_test=job.is_test
+WHERE q.mailing_id IS NULL');
     }
     catch (\Civi\Core\Exception\DBQueryException $e) {
       throw new CRM_Core_Exception(