@totten 2023-09-26
For that benchmark (50m row over 19min), that's s 2.6m rows per minute. It sounds like we could make the `UPDATE` step "safe" for web-upgrades on large DBs by setting a page-size around 500k or 1m rows.