@eileenmcnaughton 2023-09-26
@totten yeah - but also if you care about the 20 minutes for the upgrade you will probably already be pre-upgrading the tables for the alter statements. I did put it on https://lab.civicrm.org/documentation/docs/sysadmin/-/blob/master/docs/upgrade/version-specific.md

I still want to keep working on the staged upgrade but my priority is to get an upgrade that works merged for this & the other steps & then if we can iterate it all the better.