@totten 2023-09-26
Did some `r-run` against some sample data, and it successfully backfilled `mailing_id` and `is_test`.